+++ sharktank/sharktank/tools/models.json
@@ -62,8 +62,10 @@
       "sharktank"
     ],
     "output_dir": "../shark-ai/output_artifacts/",
-    "prefill_gold": 205.611,
-    "decode_gold": 8.742,
+    "prefill_gold_mi325x": 205.611,
+    "decode_gold_mi325x": 8.742,
+    "prefill_gold_mi300x": 233.195,
+    "decode_gold_mi300x": 9.716,
     "extra_compile_flags_list": [],
     "extra_benchmark_flags_list": [
       "--hip_use_streams=true"
@@ -137,8 +139,10 @@
       "sharktank"
     ],
     "output_dir": "../shark-ai/output_artifacts/",
-    "prefill_gold": 205.611,
-    "decode_gold": 8.742,
+    "prefill_gold_mi325x": 205.611,
+    "decode_gold_mi325x": 8.742,
+    "prefill_gold_mi300x": 233.257,
+    "decode_gold_mi300x": 9.680,
     "extra_compile_flags_list": [],
     "extra_benchmark_flags_list": [
       "--hip_use_streams=true"
@@ -207,8 +211,10 @@
     "bs_decode": 4,
     "extra_export_flags_list": [],
     "output_dir": "../shark-ai/output_artifacts/",
-    "prefill_gold": 2199.093,
-    "decode_gold": 74.062,
+    "prefill_gold_mi325x": 2199.093,
+    "decode_gold_mi325x": 74.062,
+    "prefill_gold_mi300x": 2646.627,
+    "decode_gold_mi300x": 83.084,
     "extra_compile_flags_list": [],
     "extra_benchmark_flags_list": [
       "--hip_use_streams=true"
@@ -282,8 +288,10 @@
       "--kv-cache-dtype=float8_e4m3fnuz"
     ],
     "output_dir": "../shark-ai/output_artifacts/",
-    "prefill_gold": 1600.852,
-    "decode_gold": 44.201,
+    "prefill_gold_mi325x": 1600.852,
+    "decode_gold_mi325x": 44.201,
+    "prefill_gold_mi300x": 1810.804,
+    "decode_gold_mi300x": 50.232,
     "extra_compile_flags_list": [],
     "extra_benchmark_flags_list": [
       "--hip_use_streams=true"
@@ -398,8 +406,10 @@
       "--kv-cache-dtype=float8_e4m3fnuz"
     ],
     "output_dir": "../shark-ai/output_artifacts/",
-    "prefill_gold": 242.228,
-    "decode_gold": 17.201,
+    "prefill_gold_mi325x": 242.228,
+    "decode_gold_mi325x": 17.201,
+    "prefill_gold_mi300x": 266.878,
+    "decode_gold_mi300x": 16.853,
     "extra_compile_flags_list": [],
     "extra_benchmark_flags_list": [
       "--device_allocator=caching"
